extract common parts of splitter based managers
