+++ tests/test_core/test_area/test_new_batch.py
@@ -57,7 +57,7 @@ def request_mock_setup(requests_mock):
 def area_config_fixture():
     return {
         "manager": "eogrow.core.area.NewBatchAreaManager",
-        "area_of_interest": {
+        "area": {
             "filename": "test_large_area.geojson",
             "buffer": 2,
             "simplification_factor": 0.1,
