+++ tests/test_core/test_area/test_new_utm.py
@@ -15,7 +15,7 @@ pytestmark = pytest.mark.fast
 @pytest.fixture(scope="session", name="large_area_config")
 def large_area_config_fixture():
     return {
-        "area_of_interest": {"filename": "test_large_area.geojson", "buffer": 1},
+        "area": {"filename": "test_large_area.geojson", "buffer": 1},
         "patch": {"size_x": 1000000, "size_y": 1000000, "buffer_x": 0, "buffer_y": 0},
     }
 
@@ -23,7 +23,7 @@ def large_area_config_fixture():
 @pytest.fixture(scope="session", name="area_config")
 def area_config_fixture():
     return {
-        "area_of_interest": {"filename": "test_area.geojson", "buffer": 0.001},
+        "area": {"filename": "test_area.geojson", "buffer": 0.001},
         "patch": {"size_x": 2400, "size_y": 1100, "buffer_x": 120, "buffer_y": 55},
     }
 
@@ -32,7 +32,7 @@ def area_config_fixture():
     "simplification_factor,expected_point_count", [(0, 128), (0.00001, 64), (0.0001, 25), (0.001, 10), (0.1, 5)]
 )
 def test_get_area_geometry(area_config, storage, simplification_factor, expected_point_count):
-    area_config["area_of_interest"]["simplification_factor"] = simplification_factor
+    area_config["area"]["simplification_factor"] = simplification_factor
     area_manager = NewUtmZoneAreaManager.from_raw_config(area_config, storage)
 
     geometry = area_manager.get_area_geometry()
